refactor: rename custom client into Http client holder and make create http client internal private function
