+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/repository/ContactsRepositoryImpl.kt
@@ -6,6 +6,7 @@ import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import com.bilalazzam.contacts_provider.ContactField
 import com.bilalazzam.contacts_provider.ContactsProvider
+import io.ktor.client.HttpClient
 import io.ktor.client.request.get
 import io.ktor.client.request.parameter
 import io.ktor.client.request.post
@@ -26,15 +27,16 @@ import net.thechance.mena.core_chat.domain.model.PagedData
 import net.thechance.mena.core_chat.domain.repository.ContactsRepository
 import kotlin.uuid.ExperimentalUuidApi
 import net.thechance.mena.core_chat.data.source.local.datastore.tryCall
-import net.thechance.mena.core_chat.data.source.remote.network.CustomHttpClient
+import net.thechance.mena.core_chat.data.source.remote.network.HttpClientHolder
 
 class ContactsRepositoryImpl(
-    private val customHttpClient: CustomHttpClient,
+    private val clientHolder: HttpClientHolder,
     private val contactsProvider: ContactsProvider,
     private val dataStore: DataStore<Preferences>
 ) : ContactsRepository{
 
-    private val client = customHttpClient.getClient()
+    private val client: HttpClient
+        get() = clientHolder.getClient()
 
     @OptIn(ExperimentalUuidApi::class)
     override suspend fun getUserContacts(pageNumber: Int): PagedData<Contact> {
