+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/repository/MessageRepositoryImpl.kt
@@ -2,6 +2,7 @@
 
 package net.thechance.mena.core_chat.data.repository
 
+import io.ktor.client.HttpClient
 import io.ktor.client.request.get
 import io.ktor.client.request.parameter
 import io.ktor.util.reflect.typeInfo
@@ -37,7 +38,7 @@ import net.thechance.mena.core_chat.data.source.remote.mapper.toListOfMessages
 import net.thechance.mena.core_chat.data.source.remote.mapper.toLocalDto
 import net.thechance.mena.core_chat.data.source.remote.mapper.toPagedListOfMessages
 import net.thechance.mena.core_chat.data.source.remote.mapper.toPendingMessageLocalDto
-import net.thechance.mena.core_chat.data.source.remote.network.CustomHttpClient
+import net.thechance.mena.core_chat.data.source.remote.network.HttpClientHolder
 import net.thechance.mena.core_chat.data.source.remote.network.WebSocketManager
 import net.thechance.mena.core_chat.data.source.remote.network.tryNetworkCall
 import net.thechance.mena.core_chat.domain.entity.Message
@@ -60,7 +61,7 @@ import kotlin.uuid.Uuid
 
 @OptIn(ExperimentalUuidApi::class)
 class MessageRepositoryImpl(
-    private val customHttpClient: CustomHttpClient,
+    private val clientHolder: HttpClientHolder,
     private val webSocketManager: WebSocketManager,
     private val pendingMessageDao: PendingMessageDao,
     private val cachedMessageDao: CachedMessageDao,
@@ -70,7 +71,8 @@ class MessageRepositoryImpl(
     private val messageSenderFactory: MessageSenderFactory,
     private val json: Json,
 ) : MessageRepository {
-    private val client = customHttpClient.getClient()
+    private val client: HttpClient
+        get() = clientHolder.getClient()
     private val messagesFlow = MutableSharedFlow<Message>()
     private val markMessagesAsRead = MutableSharedFlow<MarkMessageAsReadEvent>()
     private val markChatAsDeleted = MutableSharedFlow<DeleteChatEvent>()
