+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/repository/UserRepositoryImpl.kt
@@ -1,19 +1,21 @@
 package net.thechance.mena.core_chat.data.repository
 
+import io.ktor.client.HttpClient
 import io.ktor.client.request.get
 import io.ktor.util.reflect.typeInfo
 import net.thechance.mena.core_chat.data.source.remote.dto.UserDto
 import net.thechance.mena.core_chat.data.source.remote.mapper.toDomain
-import net.thechance.mena.core_chat.data.source.remote.network.CustomHttpClient
+import net.thechance.mena.core_chat.data.source.remote.network.HttpClientHolder
 import net.thechance.mena.core_chat.data.source.remote.network.tryNetworkCall
 import net.thechance.mena.core_chat.domain.entity.User
 import net.thechance.mena.core_chat.domain.repository.UserRepository
 
 class UserRepositoryImpl(
-    private val customHttpClient: CustomHttpClient,
+    private val clientHolder: HttpClientHolder,
 ) : UserRepository {
 
-    private val client = customHttpClient.getClient()
+    private val client: HttpClient
+        get() = clientHolder.getClient()
     override suspend fun getUserInfo(): User {
         return tryNetworkCall<UserDto>(
             bodyType = typeInfo<UserDto>()
